+++ HISTORY.md
@@ -13,10 +13,14 @@
 * Updated `get_projects_by_id()` to allow string values (URL slugs) for `project_id`
 * Updated `get_user_by_id()` to allow string values (usernames) for `user_id`
 
-### Other Changes
-* Dropped support for python 3.6
+### Authentication
 * Added support for JWT authentication, which will now be used by default
   * To get an OAuth access token instead of a JWT, call `get_access_token(jwt=False)`
+* Added caching to `get_access_token()`. JWTs will be stored in the API response cache and reused
+  until they expire.
+
+### Other Changes
+* Dropped support for python 3.6
 * Added support for setting timeout for individual API requests (with `timeout` parameter)
 * Added support for setting cache timeout for individual API requests (with `expire_after` parameter)